Bound the send-queue privilege recheck loop; split timer fire-gate logic out
throttledTwitchSend's privilege recheck loop ran before send()'s own
10s timeout ever applied, so a status that kept flipping between
checks (tmi.js userstate churn during a busy Shared Chat session, for
example) could spin it forever — wedging the single global send queue
for every later send, across every channel and feature, with no error
and nothing to log. Cap it at a fixed number of rechecks and proceed
with the last-observed status instead.

Also extracts the timer scheduler's fire-gate evaluation and
block-reason logging (added in the previous commit) into their own
module, timerCommandFireGate.ts, to keep timerCommandScheduler.ts's
complexity down now that it covers both the per-row gate and the
shared-chat cooldown/dispatch concerns.
