+++ src/twitch/twitchSendQueue.test.ts
@@ -73,6 +73,20 @@ describe('throttledTwitchSend', () => {
     expect(second).toHaveBeenCalledTimes(1);
   });
 
+  it('gives up re-checking a privileged status that never stabilizes, instead of spinning forever', async () => {
+    // A status that flips on every single check (e.g. tmi.js userstate churning mid-session)
+    // would previously spin the recheck loop indefinitely, since that loop runs before send()'s
+    // own timeout ever applies — wedging the shared queue for every later send, forever, with no
+    // error. It must give up after a bounded number of attempts and still call send().
+    let calls = 0;
+    const isPrivileged = (): boolean => { calls += 1; return calls % 2 === 0; };
+    const send = vi.fn().mockResolvedValue(undefined);
+
+    await throttledTwitchSend('streamer', isPrivileged, send);
+
+    expect(send).toHaveBeenCalledTimes(1);
+  });
+
   // ─── Shared 30s window ──────────────────────────────────────────────────────
 
   it('lets a non-privileged send run up to the 20-message limit without delay (spread across channels to dodge the per-channel floor)', async () => {
